# swagger.json
{
          "openapi": "3.1.0",
          "info": {
                    "title": "FastAPI Boilerplate",
                    "description": "FastAPI Boilerplate",
                    "version": "1.0.0"
          },
          "paths": {
...
                    "/api/v1/roles/{user_role_id}/permissions": {
                              "get": {
                                        "tags": [
                                                  "Roles"
                                        ],
                                        "summary": "Get User Role Permissions",
                                        "operationId": "get_user_role_permissions_api_v1_roles__user_role_id__permissions_get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_role_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Role Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/permissions": {
                              "get": {
                                        "tags": [
                                                  "Permissions"
                                        ],
                                        "summary": "Get Permissions",
                                        "operationId": "get_permissions_api_v1_permissions_get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "page_num",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "default": 1,
                                                                      "title": "Page Num"
                                                            }
                                                  },
                                                  {
                                                            "name": "page_size",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "default": 10,
                                                                      "title": "Page Size"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              },
                              "post": {
                                        "tags": [
                                                  "Permissions"
                                        ],
                                        "summary": "Create Permission",
                                        "operationId": "create_permission_api_v1_permissions_post",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "requestBody": {
                                                  "required": true,
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/UserPermissionCreateRequest"
                                                                      }
                                                            }
                                                  }
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/permissions/{permission_id}": {
                              "get": {
                                        "tags": [
                                                  "Permissions"
                                        ],
                                        "summary": "Get Permission",
                                        "operationId": "get_permission_api_v1_permissions__permission_id__get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "permission_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "Permission Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              },
                              "put": {
                                        "tags": [
                                                  "Permissions"
                                        ],
                                        "summary": "Update Permission",
                                        "operationId": "update_permission_api_v1_permissions__permission_id__put",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "permission_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "Permission Id"
                                                            }
                                                  }
                                        ],
                                        "requestBody": {
                                                  "required": true,
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/UserPermissionUpdateRequest"
                                                                      }
                                                            }
                                                  }
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              },
                              "delete": {
                                        "tags": [
                                                  "Permissions"
                                        ],
                                        "summary": "Delete Permission",
                                        "operationId": "delete_permission_api_v1_permissions__permission_id__delete",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "permission_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "Permission Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
...
          },
          "components": {
                    "schemas": {
...
                              "HTTPValidationError": {
                                        "properties": {
                                                  "detail": {
                                                            "items": {
                                                                      "$ref": "#/components/schemas/ValidationError"
                                                            },
                                                            "type": "array",
                                                            "title": "Detail"
                                                  }
                                        },
                                        "type": "object",
                                        "title": "HTTPValidationError"
                              },
...
                              "UserPermissionCreateRequest": {
                                        "properties": {
                                                  "permission_name": {
                                                            "type": "string",
                                                            "title": "Permission Name",
                                                            "description": "name of the permission",
                                                            "examples": [
                                                                      "create_user",
                                                                      "update_user",
                                                                      "delete_user"
                                                            ]
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "permission_name"
                                        ],
                                        "title": "UserPermissionCreateRequest"
                              },
                              "UserPermissionUpdateRequest": {
                                        "properties": {
                                                  "permission_name": {
                                                            "type": "string",
                                                            "title": "Permission Name",
                                                            "description": "name of the permission",
                                                            "examples": [
                                                                      "create_user",
                                                                      "update_user",
                                                                      "delete_user"
                                                            ]
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "permission_name"
                                        ],
                                        "title": "UserPermissionUpdateRequest"
                              },
...
                              "ValidationError": {
                                        "properties": {
                                                  "loc": {
                                                            "items": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "string"
                                                                                },
                                                                                {
                                                                                          "type": "integer"
                                                                                }
                                                                      ]
                                                            },
                                                            "type": "array",
                                                            "title": "Location"
                                                  },
                                                  "msg": {
                                                            "type": "string",
                                                            "title": "Message"
                                                  },
                                                  "type": {
                                                            "type": "string",
                                                            "title": "Error Type"
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "loc",
                                                  "msg",
                                                  "type"
                                        ],
                                        "title": "ValidationError"
                              }
                    },
                    "securitySchemes": {
                              "HTTPBearer": {
                                        "type": "http",
                                        "scheme": "bearer"
                              }
                    }
          }
}